requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohappyeyeballs=2.4.3
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.4.1=pyhd8ed1ab_1
annotated-types=0.7.0=pypi_0
anyio=4.6.2.post1=pypi_0
appnope=0.1.4=pyhd8ed1ab_0
asgiref=3.8.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pyh71513ae_0
aws-c-auth=0.7.31=h0ae3fd4_4
aws-c-cal=0.7.4=h65c637f_3
aws-c-common=0.9.30=ha44c9a9_0
aws-c-compression=0.2.19=h65c637f_3
aws-c-event-stream=0.4.3=h352a8e2_6
aws-c-http=0.8.10=hde7ce57_4
aws-c-io=0.14.20=h7f20503_1
aws-c-mqtt=0.10.7=h0c563ea_4
aws-c-s3=0.6.7=hff1d21c_2
aws-c-sdkutils=0.1.19=h65c637f_5
aws-checksums=0.1.20=h65c637f_2
aws-crt-cpp=0.28.5=h93c3e88_2
aws-sdk-cpp=1.11.407=h04e443d_3
azure-core-cpp=1.14.0=h9a36307_0
azure-identity-cpp=1.10.0=ha4e2ba9_0
azure-storage-blobs-cpp=12.13.0=h3d2f5f1_1
azure-storage-common-cpp=12.8.0=h1ccc5ac_1
azure-storage-files-datalake-cpp=12.12.0=h86941f0_1
backoff=2.2.1=pypi_0
bcrypt=4.2.0=pypi_0
blinker=1.8.2=pyhd8ed1ab_0
brotli-python=1.1.0=py39h7c0e7c0_2
build=1.2.2.post1=pypi_0
bzip2=1.0.8=hfdf4475_7
c-ares=1.34.2=h32b1619_0
ca-certificates=2024.8.30=h8857fd0_0
cachetools=5.5.0=pyhd8ed1ab_0
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py39h8ddeee6_0
charset-normalizer=3.4.0=pyhd8ed1ab_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.15=pypi_0
click=8.1.7=unix_pyh707e725_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.8.7=py39h7c0e7c0_0
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pypi_0
distro=1.9.0=pypi_0
durationpy=0.9=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
fastapi=0.115.3=pypi_0
filelock=3.16.1=pypi_0
flatbuffers=24.3.25=pypi_0
freetype=2.12.1=h60636b9_2
frozenlist=1.4.1=pypi_0
fsspec=2024.10.0=pypi_0
gflags=2.2.2=hac325c4_1005
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.43=pyhd8ed1ab_0
glog=0.7.1=h2790a97_0
google-auth=2.35.0=pypi_0
googleapis-common-protos=1.65.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.67.0=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=py39h6e9494a_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.6=pypi_0
httptools=0.6.4=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.26.1=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=75.1=h120a0e1_0
idna=3.10=pyhd8ed1ab_0
importlib-metadata=7.2.1=pyha770c72_0
importlib_resources=6.4.5=pyhd8ed1ab_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.18.1=pyh707e725_3
ipywidgets=8.1.5=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
jiter=0.6.1=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pyhd8ed1ab_0
jsonschema-specifications=2024.10.1=pyhd8ed1ab_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
jupyterlab_widgets=3.0.13=pyhd8ed1ab_0
krb5=1.21.3=h37d8d59_0
kubernetes=31.0.0=pypi_0
langchain=0.3.4=pypi_0
langchain-community=0.3.3=pypi_0
langchain-core=0.3.12=pypi_0
langchain-openai=0.2.3=pypi_0
langchain-text-splitters=0.3.0=pypi_0
langsmith=0.1.137=pypi_0
lcms2=2.16=ha2f27b4_0
lerc=4.0.0=hb486fe8_0
libabseil=20240722.0=cxx17_hac325c4_1
libarrow=17.0.0=h0c1ea87_23_cpu
libarrow-acero=17.0.0=h240833e_23_cpu
libarrow-dataset=17.0.0=h240833e_23_cpu
libarrow-substrait=17.0.0=hdefb866_23_cpu
libblas=3.9.0=25_osx64_openblas
libbrotlicommon=1.1.0=h00291cd_2
libbrotlidec=1.1.0=h00291cd_2
libbrotlienc=1.1.0=h00291cd_2
libcblas=3.9.0=25_osx64_openblas
libcrc32c=1.1.2=he49afe7_0
libcurl=8.10.1=h58e7537_0
libcxx=19.1.2=hf95d169_0
libdeflate=1.22=h00291cd_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=h10d778d_2
libevent=2.1.12=ha90c15b_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=13_2_0_h97931a8_3
libgfortran5=13.2.0=h2873a65_3
libgoogle-cloud=2.30.0=hade041e_0
libgoogle-cloud-storage=2.30.0=h8126ed0_0
libgrpc=1.65.5=hb88832f_0
libiconv=1.17=hd75f5a5_2
libjpeg-turbo=3.0.0=h0dc2134_1
liblapack=3.9.0=25_osx64_openblas
libnghttp2=1.64.0=hc7306c3_0
libopenblas=0.3.28=openmp_h8869122_0
libparquet=17.0.0=hc957f30_23_cpu
libpng=1.6.44=h4b8f8c9_0
libprotobuf=5.27.5=h62b0dff_2
libre2-11=2024.07.02=hd530cb8_1
libsodium=1.0.20=hfdf4475_0
libsqlite=3.47.0=h2f8c449_0
libssh2=1.11.0=hd019ec5_0
libthrift=0.21.0=h75589b3_0
libtiff=4.7.0=h583c2ba_1
libutf8proc=2.8.0=hb7f2c08_0
libwebp-base=1.4.0=h10d778d_0
libxcb=1.17.0=hf1f96e2_0
libxml2=2.12.7=heaf3512_4
libzlib=1.3.1=hd23fc13_2
llvm-openmp=19.1.2=hf78d878_0
lz4-c=1.9.4=hf0c8a7f_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=3.0.2=py39h20cc651_0
marshmallow=3.23.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pyhd8ed1ab_0
mmh3=5.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=1.10.0=pyhd8ed1ab_0
ncurses=6.5=hf036a51_1
nest-asyncio=1.6.0=pyhd8ed1ab_0
numpy=1.26.4=py39h28c39a1_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.19.2=pypi_0
openai=1.52.1=pypi_0
openjpeg=2.5.2=h7310d3a_0
openssl=3.3.2=hd23fc13_0
opentelemetry-api=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.27.0=pypi_0
opentelemetry-instrumentation=0.48b0=pypi_0
opentelemetry-instrumentation-asgi=0.48b0=pypi_0
opentelemetry-instrumentation-fastapi=0.48b0=pypi_0
opentelemetry-proto=1.27.0=pypi_0
opentelemetry-sdk=1.27.0=pypi_0
opentelemetry-semantic-conventions=0.48b0=pypi_0
opentelemetry-util-http=0.48b0=pypi_0
orc=2.0.2=h52ea4d3_1
orjson=3.10.10=pypi_0
overrides=7.7.0=pypi_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=py39h88a5ddd_1
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py39h25798ad_1
pip=24.2=pyh8b19718_1
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.3.6=pyhd8ed1ab_0
posthog=3.7.0=pypi_0
prompt-toolkit=3.0.48=pyha770c72_0
propcache=0.2.0=pypi_0
protobuf=4.25.5=pypi_0
psutil=6.0.0=py39h296a897_2
pthread-stubs=0.4=h00291cd_1002
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyarrow=17.0.0=py39h810aed5_1
pyarrow-core=17.0.0=py39ha39a2de_1_cpu
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycparser=2.22=pyhd8ed1ab_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydantic-settings=2.6.0=pypi_0
pydeck=0.8.0=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.20=hf24efe3_1_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pyhd8ed1ab_0
python-tzdata=2024.2=pyhd8ed1ab_0
python_abi=3.9=5_cp39
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.2=py39h06d86d0_1
pyzmq=26.2.0=py39h7644d4c_3
re2=2024.07.02=h2fb0a26_1
readline=8.2=h9e318b2_1
referencing=0.35.1=pyhd8ed1ab_0
regex=2024.9.11=pypi_0
requests=2.32.3=pyhd8ed1ab_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.3=pyhd8ed1ab_0
rpds-py=0.20.0=py39h6e5e23e_1
rsa=4.9=pypi_0
setuptools=75.1.0=pyhd8ed1ab_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pyhd8ed1ab_0
snappy=1.2.1=he1e6707_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.36=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.41.0=pypi_0
streamlit=1.39.0=pyhd8ed1ab_0
sympy=1.13.3=pypi_0
tenacity=9.0.0=pyhd8ed1ab_0
tiktoken=0.8.0=pypi_0
tk=8.6.13=h1abcd95_1
tokenizers=0.20.1=pypi_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.2=pypi_0
tornado=6.4.1=py39h06d86d0_1
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typer=0.12.5=pypi_0
typing-extensions=4.12.2=hd8ed1ab_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
tzlocal=5.2=py39h6e9494a_1
urllib3=2.2.3=pyhd8ed1ab_0
uvicorn=0.32.0=pypi_0
uvloop=0.21.0=pypi_0
validators=0.34.0=pyhd8ed1ab_0
watchdog=5.0.3=py39h06d86d0_0
watchfiles=0.24.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websocket-client=1.8.0=pypi_0
websockets=13.1=pypi_0
wheel=0.44.0=pyhd8ed1ab_0
widgetsnbextension=4.0.13=pyhd8ed1ab_0
wrapt=1.16.0=pypi_0
xorg-libxau=1.0.11=h00291cd_1
xorg-libxdmcp=1.1.5=h00291cd_0
xz=5.2.6=h775f41a_0
yaml=0.2.5=h0d85af4_2
yarl=1.16.0=pypi_0
zeromq=4.3.5=he4ceba3_6
zipp=3.20.2=pyhd8ed1ab_0
zstandard=0.23.0=py39hc23f734_1
zstd=1.5.6=h915ae27_0
